refactor: MVC 패턴에 맞게 패키지를 나눈다.
