+++ src/main/kotlin/week1/racingcar/controller/RaceController.kt
@@ -1,0 +1,14 @@
+package week1.racingcar.controller
+
+import week1.racingcar.domain.RacingGame
+
+class RaceController {
+
+    fun startRacing() {
+    }
+}
+
+fun main() {
+    RacingGame().startGame()
+    RaceController()
+}
